prototype of split CStruct macro
This commit adjusts the [CStruct] macro to take place over two stages,
one declaring the relevant [Manifest] (for the struct definition), and
the other defining operations on [Rep] to emit the correct C code.
Unfortunately, this currently requires duplicating all field definitions
between the original struct definition and the decorated [Ops] class,
which will hopefully be addressed in the future.